refactor(policingice): extract resolveVideoUrl to shared utility
Move Twitter /i/status/ URL resolution to video-utils.ts and apply it
in all video creation paths:
- Share route (web share target)
- Create incident (public submission)
- Add video to incident
- Admin bulk create
